+++ Collection/Resource/Collection.entitlements
