+++ Collection/Utility/ThumbnailProvider.swift
